sync timestamps to binance server time
There is an issue when using this library on different platforms. It doesn't sync to the server properly causing any api calls to fail since the epoch seconds are more than a second apart. This implementation attempts to sync the server and the api client as much as possible. There will be some small time between the return on the server time call and setting the offset for the synced time.